# Avoid double JSON parse for metadata-less messages