--- a/activesupport/lib/active_support/messages/metadata.rb
+++ b/activesupport/lib/active_support/messages/metadata.rb
@@ -38,11 +38,11 @@ module ActiveSupport
 
           def extract_metadata(message)
             begin
-              data = JSON.decode(message) if message.start_with?("{")
+              data = JSON.decode(message) if message.start_with?('{"_rails":')
             rescue ::JSON::JSONError
             end
 
-            if data.is_a?(Hash) && data.key?("_rails")
+            if data
               new(decode(data["_rails"]["message"]), data["_rails"]["exp"], data["_rails"]["pur"])
             else
               new(message)